Support container IAM policy
The resulting permissions are the intersection of the policy and the
role associated with the container. The result is that a new role is not
necessary to get custom permissions, which is an operational and
security problem. It also means the host machine can provide a broad
role that sets the permissions limits and the containers can operate
within those bounds.